Fix signal/slot connection for KNetworkByteStream based class.
Make Jabber plugin works.


svn path=/trunk/KDE/kdenetwork/kopete/; revision=520709

+++ protocols/yahoo/libkyahoo/yahoobytestream.h
@@ -54,7 +54,7 @@ private slots:
 	void slotConnected ();
 	void slotConnectionClosed ();
 	void slotReadyRead ();
-	void slotBytesWritten ( int );
+	void slotBytesWritten ( qint64 );
 	void slotError ( int );
 
 private:
